Begone with pkg/testutils
Literally only one file is using it, just move the parts that are used to
the package where it's used.
